Cascade exercise fix
I used the CSS methods: Cascade, Specificity and Inheritance.
This is basically to learn what the cascade does, how specificity and combining CSS selectors and how is the inheritance affects certain properties.There are multiple ways to solve it, in this ocassion i choose to do it by re-order a few rules and combining selectors.